libs/requirements.txt
boto3==1.16.36
fastjsonschema==2.14.5
google-cloud-pubsub==1.7.0
google-cloud-storage==1.30.0
pandas==1.1.5
